+++ backend/src/modules/poItem/poItem.zod.js
@@ -3,7 +3,7 @@ import { z } from "zod";
 export const createPoItemSchema = z.object({
   poId: z.string().uuid("Invalid PO ID format"),
   itemId: z.string().uuid("Invalid item ID format"),
-  quantity: z.number().int().positive("Quantity must be positive"),
+  quantity: z.number().positive("Quantity must be positive"),
   rate: z.number().positive("Rate must be positive"),
   total: z.number().positive("Total must be positive"),
 });
